python/requirements.txt
# Market Trend AI - Python Dependencies

## Core
python-dotenv>=1.0.0
requests>=2.31.0
Flask>=3.0.0
Flask-CORS>=4.0.0

## Data Processing
numpy>=1.24.0

## AI/ML (Optional - for Hugging Face transformers)
transformers>=4.35.0
torch>=2.0.0
tokenizers>=0.15.0

## Caching (Optional - for Redis support)
redis>=5.0.0

## Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

## Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
isort>=5.12.0
